+++ src/formulafence/models.py
@@ -3133,6 +3133,7 @@ class FormulaCachedResultEntry:
     result_type: str
     result_signature: str = field(repr=False)
     is_broken_reference_error: bool = False
+    is_divide_by_zero_error: bool = False
 
     @property
     def location(self) -> CellKey: